+++ src/app/(pages)/(albaform)/component/ApplyHeader.tsx
